Move some json gui objects to their own files.
